Change sketch folder name to match sketch name
The Arduino IDE requires that the folder name match the sketch name. When opening a .ino file that does not meet this requirement a dialog is displayed:

The file "smoke.ino" needs to be inside a sketch folder named "smoke".
Create this folder, move the file, and continue?

which is not the ideal initial user experience.